requirements.txt
alembic==1.7.5
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flask-marshmallow==0.14.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.31
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
requests==2.27.1
marshmallow-sqlalchemy==0.28.0
pyOpenSSL==22.0.0
email_validator==1.2.0
